requirements.txt
annotated-types==0.6.0
blinker==1.6.3
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cors==1.0.1
cryptography==41.0.4
filelock==3.12.4
firebase==4.0.1
firebase-admin==6.2.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
future==0.18.3
gevent==23.9.1
google-api-core==2.12.0
google-api-python-client==2.104.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==3.0.0
grpcio==1.59.0
grpcio-status==1.59.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
inflect==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msgpack==1.0.7
packaging==23.2
proto-plus==1.22.3
protobuf==4.24.4
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pyparsing==3.1.1
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
requests-file==1.5.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.22
style==1.1.0
tldextract==5.0.1
typing_extensions==4.8.0
update==0.0.1
uritemplate==4.1.1
urllib3==2.0.7
Werkzeug==3.0.0
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
